SCN_Gameplay

+++ Assets/Scripts/SO/EnemyInfoSO/EnemyInfoSO.cs
@@ -2,10 +2,10 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "New Enemy", menuName = "Enemy/New Enemy Info")]
 public class EnemyInfoSO : ScriptableObject
 {
-    public int maxHealth;
-    public int moveSpeed;
+    public float maxHealth;
+    public float moveSpeed;
     public int expDrop;
-    public int damage;
+    public float damage;
     public float attackRange;
     public float attackCooldown;
     public ObjectType type;
